+++ src/main/java/com/infeco/keylease/repository/AddressRepository.java
@@ -1,9 +1,11 @@
 package com.infeco.keylease.repository;
 
-import com.infeco.keylease.entity.AuthorityEntity;
+import com.infeco.keylease.entity.AddressEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.UUID;
+
 @Repository
-public interface AuthorityRepository extends JpaRepository<AuthorityEntity, String> {
+public interface AddressRepository extends JpaRepository<AddressEntity, UUID> {
 }